Refactor Interface Segregation and DI applied
