--- TheWarriors/src/main/java/com/warriors/model/warriors/Fightable.java
@@ -1,9 +1,0 @@
-package com.warriors.model.warriors;
-
-import com.warriors.model.damage.Damage;
-
-public interface Fightable {
-    boolean isAlive();
-    void hit(Fightable opponent);
-    void receiveDamage(Damage damage, Fightable damageDealer);
-}
