+++ TheWarriors/src/main/java/com/warriors/model/warriors/interfaces/Attackable.java
@@ -1,0 +1,5 @@
+package com.warriors.model.warriors.interfaces;
+
+public interface Attackable {
+    int getAttack();
+}
