+++ TheWarriors/src/main/java/com/warriors/model/warriors/interfaces/Defendable.java
@@ -1,4 +1,4 @@
-package com.warriors.model.warriors;
+package com.warriors.model.warriors.interfaces;
 
 public interface Defendable {
