+++ TheWarriors/src/main/java/com/warriors/model/warriors/interfaces/Fightable.java
@@ -1,0 +1,10 @@
+package com.warriors.model.warriors.interfaces;
+
+public interface Fightable extends Attackable, HasHealth {
+    default void hit(Fightable opponent) {
+        opponent.receiveHit(this);
+    }
+    default void receiveHit(Attackable damageDealer) {
+        reduceHealthBasedOnDamage(damageDealer.getAttack());
+    }
+}
