+++ TheWarriors/src/main/java/com/warriors/model/warriors/interfaces/HasHealth.java
@@ -1,0 +1,9 @@
+package com.warriors.model.warriors.interfaces;
+
+public interface HasHealth {
+    default boolean isAlive() {
+        return getHealth() > 0;
+    }
+    void reduceHealthBasedOnDamage(int damage);
+    int getHealth();
+}
